feat: 增加ddc服务批量操作接口 #2776
